Refactor creating package.json into it's own function / module
